Create distfeat.py
Created a representation of distinctive features as `TypedDict`s, which allow for specificiation of the feature we are giving it, but also a way to convert it to a vector.

Representing each feature value as a ternary, within the set of {-1,0,1}, where -1 := the *lack* of that feature, 0 := the *negative* of that feature, and 1 := the *positive* of that feature. Ultimately, it doesn't matter, as these will be normalized later on!